refactor: collapse trainer config path, undo legacy AgentTrainer deprecation
- Move rllm/trainer/config/unified/* up one level. Hydra search paths shift
  from pkg://rllm.trainer.config.unified to pkg://rllm.trainer.config, which
  is the same package as the existing legacy yamls — no name collisions.
- Canonicalize the unified-trainer import: every example, cookbook, doc, and
  test now uses `from rllm.trainer import AgentTrainer` (which resolves to
  the unified one). The explicit `from rllm.trainer.unified_trainer import
  AgentTrainer` form still works for callers that want it.
- Drop the DeprecationWarning + fireworks-removal on the legacy
  `rllm.trainer.agent_trainer.AgentTrainer`. It stays a verl/fireworks-only
  variant alongside the canonical unified trainer.
