+++ .github/workflows/test-gateway-manager.yml
@@ -5,14 +5,14 @@ on:
     branches: [main]
     paths:
       - "rllm/gateway/manager.py"
-      - "rllm/trainer/config/unified/rllm/base.yaml"
+      - "rllm/trainer/config/rllm/base.yaml"
       - "tests/engine/test_gateway_manager.py"
       - ".github/workflows/test-gateway-manager.yml"
   pull_request:
     branches: [main]
     paths:
       - "rllm/gateway/manager.py"
-      - "rllm/trainer/config/unified/rllm/base.yaml"
+      - "rllm/trainer/config/rllm/base.yaml"
       - "tests/engine/test_gateway_manager.py"
       - ".github/workflows/test-gateway-manager.yml"
   workflow_dispatch:
