+++ cookbooks/workflow/geo3k/train.py
@@ -13,10 +13,10 @@ from omegaconf import DictConfig
 
 from rllm.data.dataset import DatasetRegistry
 from rllm.rewards.reward_fn import math_reward_fn
-from rllm.trainer.unified_trainer import AgentTrainer
+from rllm.trainer import AgentTrainer
 
 
-@hydra.main(config_path="pkg://rllm.trainer.config.unified", config_name="unified", version_base=None)
+@hydra.main(config_path="pkg://rllm.trainer.config", config_name="unified", version_base=None)
 def main(config: DictConfig):
     train_dataset = DatasetRegistry.load_dataset("geo3k", "train")
     val_dataset = DatasetRegistry.load_dataset("geo3k", "test")
