+++ docs/backends/tinker.mdx
@@ -68,9 +68,8 @@ from math_flow import math_flow                # from cookbooks/math/
 from math_eval import math_evaluator           # from cookbooks/math/
 
 from rllm.data.dataset import DatasetRegistry
-from rllm.trainer.unified_trainer import AgentTrainer
-
-@hydra.main(config_path="pkg://rllm.trainer.config.unified", config_name="unified", version_base=None)
+from rllm.trainer import AgentTrainer
+@hydra.main(config_path="pkg://rllm.trainer.config", config_name="unified", version_base=None)
 def main(config: DictConfig):
     train_dataset = DatasetRegistry.load_dataset("gsm8k", "train")
     test_dataset = DatasetRegistry.load_dataset("math500", "test")
@@ -112,11 +111,11 @@ from omegaconf import DictConfig
 from solver_judge_flow import SolverJudgeWorkflow
 
 from rllm.data.dataset import DatasetRegistry
-from rllm.trainer.unified_trainer import AgentTrainer
+from rllm.trainer import AgentTrainer
 from rllm.rewards.countdown_reward import countdown_reward_fn
 
 
-@hydra.main(config_path="pkg://rllm.trainer.config.unified", config_name="unified", version_base=None)
+@hydra.main(config_path="pkg://rllm.trainer.config", config_name="unified", version_base=None)
 def main(config: DictConfig):
     train_dataset = DatasetRegistry.load_dataset("countdown", "train")
     val_dataset = DatasetRegistry.load_dataset("countdown", "test")
