+++ docs/tutorials/solver-judge-workflow.mdx
@@ -386,10 +386,8 @@ from omegaconf import DictConfig
 from solver_judge_flow import solver_judge_flow
 
 from rllm.data.dataset import DatasetRegistry
-from rllm.trainer.unified_trainer import AgentTrainer
-
-
-@hydra.main(config_path="pkg://rllm.trainer.config.unified", config_name="unified", version_base=None)
+from rllm.trainer import AgentTrainer
+@hydra.main(config_path="pkg://rllm.trainer.config", config_name="unified", version_base=None)
 def main(config: DictConfig):
     train_dataset = DatasetRegistry.load_dataset("countdown", "train")
     test_dataset = DatasetRegistry.load_dataset("countdown", "test")
