+++ examples/agentcore_math/train_agentcore_math_tinker.py
@@ -19,7 +19,7 @@ from rllm.trainer.tinker.tinker_backend import TinkerBackend
 from rllm.trainer.unified_trainer import UnifiedTrainer
 
 
-@hydra.main(config_path="pkg://rllm.trainer.config.unified", config_name="unified", version_base=None)
+@hydra.main(config_path="pkg://rllm.trainer.config", config_name="unified", version_base=None)
 def main(config):
     train_dataset = DatasetRegistry.load_dataset("gsm8k_agentcore", "train")
     test_dataset = DatasetRegistry.load_dataset("gsm8k_agentcore", "test")
