+++ examples/harbor_swe/train_harbor.py
@@ -1,13 +1,13 @@
 import hydra
 
 from rllm.data.dataset import DatasetRegistry
-from rllm.trainer.unified_trainer import AgentTrainer
+from rllm.trainer import AgentTrainer
 
 TRAIN_DATASET = "swesmith"
 VAL_DATASET = "swebench-verified"
 
 
-@hydra.main(config_path="pkg://rllm.trainer.config.unified", config_name="unified", version_base=None)
+@hydra.main(config_path="pkg://rllm.trainer.config", config_name="unified", version_base=None)
 def main(config):
     train_dataset = DatasetRegistry.load_dataset(TRAIN_DATASET, "default")
     val_dataset = DatasetRegistry.load_dataset(VAL_DATASET, "default")
